+++ app/src/main/java/protect/rentalcalc/PropertyWorksheetActivity.java
@@ -68,7 +68,7 @@ public class PropertyWorksheetActivity extends AppCompatActivity
             return;
         }
 
-        final int propertyId = b.getInt("id");
+        final long propertyId = b.getLong("id");
         _property = _db.getProperty(propertyId);
 
         if(_property == null)
